Register single method during endpoint test
